requirements.txt
# RoofNN — Telegram-бот + Mini App
# Python 3.10+

aiogram>=3.13.0
sqlalchemy>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
